+++ hive/lib/src/binary/frame.dart
@@ -3,8 +3,8 @@ import 'dart:typed_data';
 import 'package:hive/hive.dart';
 import 'package:hive/src/binary/binary_reader_impl.dart';
 import 'package:hive/src/binary/binary_writer_impl.dart';
-import 'package:hive/src/binary/crc32.dart';
 import 'package:hive/src/crypto_helper.dart';
+import 'package:hive/src/util/crc32.dart';
 
 class Frame {
   final dynamic key;