+++ hive/lib/src/box/lazy_box_impl.dart
@@ -2,7 +2,6 @@ import 'package:hive/hive.dart';
 import 'package:hive/src/backend/storage_backend.dart';
 import 'package:hive/src/binary/frame.dart';
 import 'package:hive/src/box/box_base.dart';
-import 'package:hive/src/box/box_options.dart';
 import 'package:hive/src/box/change_notifier.dart';
 import 'package:hive/src/box/keystore.dart';
 import 'package:hive/src/hive_impl.dart';
@@ -11,11 +10,11 @@ class LazyBoxImpl extends BoxBase implements LazyBox {
   LazyBoxImpl(
     HiveImpl hive,
     String name,
-    BoxOptions options,
-    StorageBackend backend, [
     Keystore keystore,
+    CompactionStrategy compactionStrategy,
+    StorageBackend backend, [
     ChangeNotifier notifier,
-  ]) : super(hive, name, options, backend, keystore, notifier);
+  ]) : super(hive, name, keystore, compactionStrategy, backend, notifier);
 
   @override
   final bool lazy = true;
@@ -95,6 +94,16 @@ class LazyBoxImpl extends BoxBase implements LazyBox {
     notifier.notify(frames);
 
     await performCompactionIfNeeded();
+  }
+
+  @override
+  List<E> listView<E>() {
+    throw UnsupportedError('Only non-lazy boxes support listView().');
+  }
+
+  @override
+  Map<dynamic, E> mapView<E>() {
+    throw UnsupportedError('Only non-lazy boxes support mapView().');
   }
 
   @override